Now ensure that the array indicies are integers.

+++ 2dmatrix/2dmatrix.h
@@ -76,7 +76,7 @@ void mtxprntfile(matrix*, char*);
 void mtxprntfilehdr(matrix*, char*, char*);
 matrix* mtxloadcsv(char*, int);
 
-#define val(MATRIX, ROW, COL) (MATRIX)->array[(ROW)][(COL)]
+#define val(MATRIX, ROW, COL) (MATRIX)->array[(int) (ROW)][(int) (COL)]
 
 #endif
